# Provide a lock-free API in ContinuousTrajectory