--- a/physics/ephemeris.hpp
+++ b/physics/ephemeris.hpp
@@ -346,6 +346,7 @@ class Ephemeris {
   Instant instance_time() const EXCLUDES(lock_);
 
   virtual Instant t_min_locked() const REQUIRES_SHARED(lock_);
+  virtual Instant t_max_locked() const REQUIRES_SHARED(lock_);
 
   // Computes the accelerations between one body, |body1| (with index |b1| in
   // the |positions| and |accelerations| arrays) and the bodies |bodies2| (with